add account and transaction repo and service and gandler and make more better struct

+++ internal/models/account.go
@@ -1,12 +1,12 @@
 package models
 
 type Account struct {
-	ID       int    `json:"id"`
-	Iban     string `json:"iban"`
-	UserID   int    `json:"user_id"`
-	Currency string `json:"currency"`
-	Blocked  bool   `json:"blocked"`
-	Amount   int    `json:"amount"`
+	ID       int     `json:"id"`
+	Iban     string  `json:"iban"`
+	UserID   int     `json:"user_id"`
+	Currency string  `json:"currency"`
+	Blocked  bool    `json:"blocked"`
+	Amount   float64 `json:"amount"`
 }
 
 type Orderings map[string]string
